Module Wikidata

type propertyid = string

Used for strings that contain Wikidata Property IDs, like "P69".

type lang = string

Used for strings that contain Wikimedia language codes.

exception Bad_json_data of string

Called when the json input is missing some field that was expected. The string describes what exactly is missing

exception Not_implemented

Called when using a feature that is not implemented. Currently, this is only raised when attempting to parse a Lexeme, Form, or Sense.

module Snak : sig ... end

Represents Wikidata Snaks.

module Statement : sig ... end

Represents Wikidata Statements.

module Entity : sig ... end

Represents Wikidata Entities.