Module Entity.Item

Represents Wikidata Items.

Sitelinks

type sitelink = {

site : string;

title : string;

badges : string list;

url : string option;

}

Links to the corresponding Wikipedia articles for this item, closely following the representation of sitelinks in the original JSON as specified in the Wikibase JSON specifications.

Items

type t = < id : string; entity_type : string; label : lang -> string; aliases : lang -> string list; description : lang -> string; claim_groups : (propertyid * Statement.t list) list; claim_group : propertyid -> Statement.t list; truthy_claim_groups : (propertyid * Statement.t list) list; truthy_claim_group : propertyid -> Statement.t list; sitelinks : (string * sitelink) list; >

Represents Wikidata Items.

id : string

Returns the QID of the item.

entity_type : string

Returns the entry for type in the original json. For Items, should always return "item"

label : lang -> string

Given a Wikidata language code, returns the label of the Item in that language.

aliases : lang -> string list

Given a Wikidata language code, returns a list of aliases for the Item in that language.

description : lang -> string

Given a Wikidata language code, returns the description for the Item in that language.

claim_groups : (propertyid * Statement.t list) list

Returns all Statements/Claims about this Item in an associative list between Property IDs and their corresponding Statements.

claim_group : propertyid -> Statement.t list

Given a Property ID, returns the list of Statements/Claims corresponding to that Property for this Item. If that Property ID isn't present, returns an empty list.

truthy_claim_groups : (propertyid * Statement.t list) list

Returns all "truthy" Statements/Claims about this Item in an associative list between Property IDs and their corresponding Statements.

truthy_claim_group : propertyid -> Statement.t list

Given a Property ID, returns the list of "truthy" Statements/Claims corresponding to that Property for this Item. If that Property ID isn't present or there are none, returns an empty list.

sitelinks : (string * sitelink) list

Returns an associative list between Wikipedia language site identifiers and sitelinks.

Conversions

val of_string : string -> t

Turns a json string representing an Item (as described in the Wikibase JSON specification) into a t.

val of_entities_string : string -> t

Returns the first entity in a JSON string representing a group of entities as a t. This is provided as a convenience function, as this is the format that requests to https://wikidata.org/wiki/Special:EntityData/{ID}.json (where ID is a Wikidata Entity ID) are returned in.