Class Item.t

Represents Wikidata Items.

Note that while this object may be constructed from this class manually, the expected use is to use the of_string and of_entities_string functions instead.

inherit basic_entity

inherit label_description_aliases_mixin

inherit statements_mixin

method sitelinks : (string * sitelink) list

Returns an associative list between Wikipedia language site identifiers and sitelinks.