Module Entity.Property

Represents Wikidata Properties.

class t : id:string -> entity_type:string -> labels:(lang * string) list -> descriptions:(lang * string) list -> aliases:(lang * string list) list -> statements:(propertyid * Statement.t list) list -> datatype:string -> object ... end

Represents Wikidata Properties.

Conversions

val of_string : string -> t

Turns a json string representing a Property (as described in the Wikibase JSON specification) into a t.

val of_entities_string : string -> t

Returns the first entity in a JSON string representing a group of entities as a t. This is provided as a convenience function, as this is the format that requests to https://wikidata.org/wiki/Special:EntityData/{ID}.json (where ID is a Wikidata Entity ID) are returned in.