Class Property.t

Represents Wikidata Properties.

Note that while this object may be constructed from this class manually, the expected use is to use the of_string and of_entities_string functions instead.

inherit basic_entity

inherit label_description_aliases_mixin

inherit statements_mixin

method datatype : string

Returns a string representing the expected datatype of its mainsnak. No guarantees are made about mappings from strings returned by this method to the variants of Snak.data.