Class Entity.label_description_aliases_mixin

Mixin for labels, descriptions, and aliases

method label : lang -> string

Given a Wikidata language code, returns the label of the Entity in that language.

method label_opt : lang -> string option

Given a Wikidata language code, returns the label of the Entity in that language as an option type, returning None if no label is found for that language.

method aliases : lang -> string list

Given a Wikidata language code, returns a list of aliases for the Entity in that language. Returns an empty list if none are found.

method description : lang -> string

Given a Wikidata language code, returns the description for the Entity in that language.

method description_opt : lang -> string option

Given a Wikidata language code, returns the description for the Entity in that language as an option type, returning None if none is found for that language.