Class Entity.statements_mixin

Mixin for Statements.

method all_statements : (propertyid * Statement.t list) list

Returns all Statements about this Entity in an associative list between Property IDs and their corresponding Statements.

method statements : propertyid -> Statement.t list

Given a Property ID, returns the list of Statements corresponding to that Property for this Entity. If that Property ID isn't present, returns an empty list.

method all_truthy_statements : (propertyid * Statement.t list) list

Returns all "truthy" Statements about this Entity in an associative list between Property IDs and their corresponding Statements.

method truthy_statements : propertyid -> Statement.t list

Given a Property ID, returns the list of "truthy" Statements corresponding to that Entity for this Item. If that Property ID isn't present or there are none, returns an empty list.