Class Item.t

Represents Wikidata Items.

Note that while this object may be constructed from this class manually, the expected use is to use the of_string and of_entities_string functions instead.

method id : string

Returns the QID of the item.

method entity_type : string

Returns the entry for type in the original json. For Items, should always return "item".

method label : lang -> string

Given a Wikidata language code, returns the label of the Item in that language.

method aliases : lang -> string list

Given a Wikidata language code, returns a list of aliases for the Item in that language.

method description : lang -> string

Given a Wikidata language code, returns the description for the Item in that language.

method all_statements : (propertyid * Statement.t list) list

Returns all Statements about this Item in an associative list between Property IDs and their corresponding Statements.

method statements : propertyid -> Statement.t list

Given a Property ID, returns the list of Statements corresponding to that Property for this Item. If that Property ID isn't present, returns an empty list.

method all_truthy_statements : (propertyid * Statement.t list) list

Returns all "truthy" Statements about this Item in an associative list between Property IDs and their corresponding Statements.

method truthy_statements : propertyid -> Statement.t list

Given a Property ID, returns the list of "truthy" Statements corresponding to that Property for this Item. If that Property ID isn't present or there are none, returns an empty list.

method sitelinks : (string * sitelink) list

Returns an associative list between Wikipedia language site identifiers and sitelinks.