Module Entity.Property

Represents Wikidata Properties.

type t = < id : string; entity_type : string; datatype : string; label : lang -> string; aliases : lang -> string list; description : lang -> string; all_statements : (propertyid * Statement.t list) list; statements : propertyid -> Statement.t list; all_truthy_statements : (propertyid * Statement.t list) list; truthy_statements : propertyid -> Statement.t list; >

Represents Wikidata Properties.

id : string

Returns the Property ID of the Property.

entity_type : string

Returns the entry for type in the original json. For Properties, should always return "property"

label : lang -> string

Given a Wikidata language code, returns the label of the Property in that language.

aliases : lang -> string list

Given a Wikidata language code, returns a list of aliases for the Property in that language.

description : lang -> string

Given a Wikidata language code, returns the description for the Property in that language.

all_statements : (propertyid * Statement.t list) list

Returns all Statements about this Property in an associative list between Property IDs and their corresponding Statements.

statements : propertyid -> Statement.t list

Given a Property ID, returns the list of Statements corresponding to that Property for this Property. If that Property ID isn't present, returns an empty list.

all_truthy_statements : (propertyid * Statement.t list) list

Returns all "truthy" Statements about this Property in an associative list between Property IDs and their corresponding Statements.

truthy_statements : propertyid -> Statement.t list

Given a Property ID, returns the list of "truthy" Statements corresponding to that Property for this Property. If that Property ID isn't present or there are none, returns an empty list.

Conversions

val of_string : string -> t

Turns a json string representing a Property (as described in the Wikibase JSON specification) into a t.

val of_entities_string : string -> t

Returns the first entity in a JSON string representing a group of entities as a t. This is provided as a convenience function, as this is the format that requests to https://wikidata.org/wiki/Special:EntityData/{ID}.json (where ID is a Wikidata Entity ID) are returned in.