Class Property.t

Represents Wikidata Properties.

Note that while this object may be constructed from this class manually, the expected use is to use the of_string and of_entities_string functions instead.

method id : string

Returns the Property ID of the Property.

method entity_type : string

Returns the entry for type in the original json. For Properties, should always return "property"

method datatype : string

Returns a string representing the expected datatype of its mainsnak. No guarantees are made about mappings from strings returned by this method to the variants of Snak.data.

method label : lang -> string

Given a Wikidata language code, returns the label of the Property in that language.

method label_opt : lang -> string option

Given a Wikidata language code, returns the label of the Property in that language as an option type, returning None if none is found for that language.

method aliases : lang -> string list

Given a Wikidata language code, returns a list of aliases for the Property in that language. Returns an empty list if none are found.

method description : lang -> string

Given a Wikidata language code, returns the description for the Property in that language.

method description_opt : lang -> string option

Given a Wikidata language code, returns the description for the Property in that language as an option type, returning None if none is found for that language.

method all_statements : (propertyid * Statement.t list) list

Returns all Statements about this Property in an associative list between Property IDs and their corresponding Statements.

method statements : propertyid -> Statement.t list

Given a Property ID, returns the list of Statements corresponding to that Property for this Property. If that Property ID isn't present, returns an empty list.

method all_truthy_statements : (propertyid * Statement.t list) list

Returns all "truthy" Statements about this Property in an associative list between Property IDs and their corresponding Statements.

method truthy_statements : propertyid -> Statement.t list

Given a Property ID, returns the list of "truthy" Statements corresponding to that Property for this Property. If that Property ID isn't present or there are none, returns an empty list.